extract override loading code
