+++ LethalCompanyInputUtils/Api/LcInputActions.cs
@@ -8,7 +8,6 @@ using LethalCompanyInputUtils.BindingPathEnums;
 using LethalCompanyInputUtils.Config;
 using LethalCompanyInputUtils.Data;
 using LethalCompanyInputUtils.Utils;
-using Newtonsoft.Json;
 using UnityEngine;
 using UnityEngine.InputSystem;
 
@@ -139,143 +138,5 @@ public abstract class LcInputActions
         File.WriteAllText(overrideType.GetJsonPath(Id), overrides.AsJson());
     }
 
-    internal void Load()
-    {
-        var overridePriority = InputUtilsConfig.bindingOverridePriority.Value;
-        
-        switch (overridePriority)
-        {
-            case BindingOverridePriority.GlobalThenLocal:
-                Load(BindingOverrideType.Local, true);
-                Load(BindingOverrideType.Global, false);
-                break;
-            case BindingOverridePriority.LocalThenGlobal:
-                Load(BindingOverrideType.Global, true);
-                Load(BindingOverrideType.Local, false);
-                break;
-            case BindingOverridePriority.GlobalOnly:
-                Load(BindingOverrideType.Global, true);
-                break;
-            case BindingOverridePriority.LocalOnly:
-                Load(BindingOverrideType.Local, true);
-                break;
-            default:
-                throw new ArgumentOutOfRangeException(nameof(overridePriority), overridePriority, null);
-        }
-    }
-
-    internal void Load(BindingOverrideType overrideType, bool removeExistingOverrides)
-    {
-        try
-        {
-            ApplyMigrations();
-        }
-        catch (Exception e)
-        {
-            Logging.Error("Got error when applying migrations, skipping...");
-            Logging.Error(e);
-        }
-
-        var jsonPath = overrideType.GetJsonPath(Id);
-
-        if (!File.Exists(jsonPath))
-            return;
-
-        try
-        {
-            var overrides = BindingOverrides.FromJson(File.ReadAllText(jsonPath));
-
-            if (removeExistingOverrides)
-                Asset.RemoveAllBindingOverrides();
-            
-            overrides.LoadInto(Asset);
-        }
-        catch (Exception e)
-        {
-            Logging.Error(e);
-        }
-    }
-
-    private void ApplyMigrations()
-    {
-        // 0.7.0 added a global overrides locations, anything beforehand should assume local path.
-        var pre070JsonPath = BindingOverrideType.Local.GetJsonPath(Id);
-        
-        // The overrides json path was changed to be nested in the config folder so that they could be packaged with modpacks.
-        var pre041JsonPath = Path.Combine(FsUtils.Pre041ControlsDir, $"{Id}.json");
-        if (File.Exists(pre041JsonPath) && !File.Exists(pre070JsonPath))
-            File.Move(pre041JsonPath, pre070JsonPath);
-
-        // If there isn't an existing overrides json then there is no need to apply any migrations
-        if (!File.Exists(pre070JsonPath))
-            return;
-
-        // Unbound binds now use a special identifier as leaving them blank prevented the overrides from actually taking effect.
-        var hasPre043Overrides = File.ReadAllText(pre070JsonPath)
-            .Replace(" ", "")
-            .Contains("\"origPath\":\"\"");
-
-        if (hasPre043Overrides)
-        {
-            var overrides = BindingOverrides.FromJson(File.ReadAllText(pre070JsonPath));
-
-            for (var i = 0; i < overrides.overrides.Count; i++)
-            {
-                var bindingOverride = overrides.overrides[i];
-                if (!string.IsNullOrEmpty(bindingOverride.origPath))
-                    continue;
-
-                if (bindingOverride.path is null)
-                    continue;
-
-                if (bindingOverride.path.StartsWith("<Keyboard>", StringComparison.InvariantCultureIgnoreCase)
-                    || bindingOverride.path.StartsWith("<Mouse>", StringComparison.InvariantCultureIgnoreCase))
-                {
-                    bindingOverride.origPath = UnboundKeyboardAndMouseIdentifier;
-                }
-                else
-                {
-                    // It's possible someone may use a unique path, such as a DualSense touchpad
-                    // so just assume that if their path wasn't either a <Keyboard> or <Mouse>, that it's probably a <Gamepad>
-                    bindingOverride.origPath = UnboundGamepadIdentifier;
-                }
-
-                overrides.overrides[i] = bindingOverride;
-            }
-
-            File.WriteAllText(pre070JsonPath, JsonConvert.SerializeObject(overrides));
-        }
-        
-        // 0.7.0 added groups to the BindingOverride schema for better device specific bind detection.
-        var hasPre070Overrides = !File.ReadAllText(pre070JsonPath)
-            .Replace(" ", "")
-            .Contains("\"groups\":");
-
-        if (hasPre070Overrides)
-        {
-            var overrides = BindingOverrides.FromJson(File.ReadAllText(pre070JsonPath));
-
-            for (int i = 0; i < overrides.overrides.Count; i++)
-            {
-                var bindingOverride = overrides.overrides[i];
-                if (string.IsNullOrEmpty(bindingOverride.origPath))
-                    continue;
-
-                if (bindingOverride.origPath.StartsWith("<Keyboard>", StringComparison.InvariantCultureIgnoreCase)
-                    || bindingOverride.origPath.StartsWith("<Mouse>", StringComparison.InvariantCultureIgnoreCase)
-                    || string.Equals(bindingOverride.origPath, UnboundKeyboardAndMouseIdentifier))
-                {
-                    bindingOverride.groups = DeviceGroups.KeyboardAndMouse;
-                }
-                else
-                {
-                    bindingOverride.groups = DeviceGroups.Gamepad;
-                }
-
-                overrides.overrides[i] = bindingOverride;
-            }
-            
-            File.WriteAllText(pre070JsonPath, JsonConvert.SerializeObject(overrides));
-        }
-    }
+    internal void Load() => Asset.LoadOverridesFromControls(Id);
 }
